moved apis into routers
