+++ Angular/backend/src/server.ts
@@ -1,7 +1,7 @@
 import express from "express";
 import cors from "cors";
-import { sample_foods, sample_tags, sample_users } from "./data";
-import jwt from "jsonwebtoken";
+import foodRouter from "./router/food.router";
+import userRouter from "./router/login.router";
 
 const app = express();
 app.use(express.json());
@@ -12,60 +12,8 @@ app.use(
 	}),
 );
 
-app.get("/api/foods", (req, res) => {
-	res.send(sample_foods);
-});
-
-app.get("/api/foods/search/:searchTerm", (req, res) => {
-	const searchTerm = req.params.searchTerm;
-	const foods = sample_foods.filter((food) =>
-		food.name.toLowerCase().includes(searchTerm.toLowerCase()),
-	);
-	res.send(foods);
-});
-
-app.get("/api/foods/tags", (req, res) => {
-	res.send(sample_tags);
-});
-
-app.get("/api/foods/tag/:tagName", (req, res) => {
-	const tagName = req.params.tagName;
-	const foods = sample_foods.filter((food) => food.tags?.includes(tagName));
-	res.send(foods);
-});
-
-app.get("/api/foods/:foodId", (req, res) => {
-	const foodId = req.params.foodId;
-	const food = sample_foods.find((food) => food.id === foodId);
-	res.send(food);
-});
-
-app.post("/api/users/login", (req, res) => {
-	const body = req.body;
-	const { email, password } = req.body;
-	const user = sample_users.find(
-		(user) => user.email === email && user.password === password,
-	);
-
-	if (user) {
-		res.send(generateTokenResponse(user));
-	} else {
-		res.status(400).send("Username or password was not found.");
-	}
-});
-
-const generateTokenResponse = (user: any) => {
-	const token = jwt.sign(
-		{ email: user.email, isAdmin: user.isAdmin },
-		"Checking...",
-		{
-			expiresIn: "30d",
-		},
-	);
-
-	user.token = token;
-	return user;
-};
+app.use("/api/foods", foodRouter);
+app.use("/api/users", userRouter);
 
 const port = 5000;
 app.listen(port, () => {
